requirements.txt
# Core dependencies
tensorflow>=2.10.0
torch>=1.12.0
transformers>=4.20.0
datasets>=2.0.0

# Experiment tracking
mlflow>=1.28.0

# Hyperparameter tuning
optuna>=3.0.0

# Model interpretability
shap>=0.41.0
lime>=0.2.0.1

# API framework
fastapi>=0.85.0
uvicorn>=0.18.0

# Data processing
pandas>=1.4.0
numpy>=1.21.0
scikit-learn>=1.1.0

# Utilities
pyyaml>=6.0
python-dotenv>=0.20.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development tools
black>=22.0.0
flake8>=5.0.0